requirements_local.txt
Flask>=3.0.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pandas>=2.1.0
openpyxl>=3.1.0
gunicorn>=21.2.0
